requirements.txt
google-generativeai==0.5.1
python-dotenv==0.19.0
streamlit==1.5.0
torch==1.10.0
transformers==4.12.2
